a suggestion to update README.md file
I think the explanation for the homework is not enough.
In the class we did not do 4 CRUD actions and in the readme.md file, the 4 crud actions are not mentioned directly. That's why at first I could not get that we had to do overall 8 actions and I have done just 4 of them.
Maybe in the readme.md file, all action is better mentioned.